implemented attribute-based event metadata store
