+++ OutboxSample/Infrastructure/AttrbiuteSourcedEventMetadataProvider.cs
@@ -1,0 +1,34 @@
+﻿using OutboxSample.Application;
+using OutboxSample.Model.Events;
+using System.Collections.Concurrent;
+
+namespace OutboxSample.Infrastructure;
+
+public class AttrbiuteSourcedEventMetadataProvider : IEventMetadataProvider
+{
+    private readonly ConcurrentDictionary<Type, EventMetadata> metadataStore;
+
+    public AttrbiuteSourcedEventMetadataProvider()
+    {
+        this.metadataStore = new ConcurrentDictionary<Type, EventMetadata>();
+    }
+
+    public EventMetadata GetMetadataFor<TEvent>(TEvent @event) where TEvent : IEvent
+    {
+        EventMetadata metadata = metadataStore.GetOrAdd(typeof(TEvent), ExtractMetadata);
+
+        return metadata;
+    }
+
+    private static EventMetadata ExtractMetadata(Type eventType)
+    {
+        var metadataAttribute = (EventMetadataAttribute?)Attribute.GetCustomAttribute(eventType, typeof(EventMetadataAttribute));
+
+        if (metadataAttribute is null)
+        {
+            throw new InvalidOperationException($"Event of type {eventType.FullName} doesn't have metadata attached.");
+        }
+
+        return new EventMetadata(metadataAttribute.EventType, metadataAttribute.AggregateType, metadataAttribute.EventSchemaVersion);
+    }
+}
